Move MetataskSealStack out of controls.tsx to avoid faction-registry import cycle
controls.tsx importing MetaTaskSeal pulled the faction registry (which statically
imports every composer) into the submit-wiring test's controls mock, binding
Coven's PublishButton to the un-mocked export. Extract the stack to its own leaf
module; repoint all 16 archetypes; convert Everymen's hand-rolled list too.
