+++ frontend/src/pages/editPraxis/archetypes/WowEditPraxis.tsx
@@ -35,11 +35,11 @@ import {
   DropButton,
   FilePicker,
   InviteSearch,
-  MetataskSealStack,
   ModePicker,
   PublishButton,
   TitleField,
 } from "./controls";
+import { MetataskSealStack } from "../MetataskSealStack";
 import type { EditPraxisState } from "../useEditPraxis";
 
 interface Props {
